requirements.txt
anyio==3.6.2
arrow==1.2.3
binaryornot==0.4.4
cashews==5.3.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
chevron==0.14.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
distlib==0.3.6
fastapi==0.92.0
filelock==3.9.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
MarkupSafe==2.1.2
msgpack==1.0.4
multidict==6.0.4
nb-cli==1.0.5
nonebot-adapter-onebot==2.2.1
nonebot-plugin-gocqhttp==0.6.7
nonebot2==2.0.0rc3
noneprompt==0.1.9
platformdirs==2.6.2
prompt-toolkit==3.0.38
psutil==5.9.4
py-cpuinfo==9.0.0
pydantic==1.10.5
pyfiglet==0.8.post1
pygtrie==2.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
text-unidecode==1.3
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
win32-setctime==1.1.0
yarl==1.8.2